Moved Note mapper and Note model to base notepad module.
