+++ base/notepad/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.library'
     id 'kotlin-android'
+    id 'kotlin-kapt'
 }
 
 apply from: rootProject.file('dependencies.gradle')
@@ -29,6 +30,10 @@ android {
 
 dependencies {
     implementation(
-            deps.kotlin.stdlib
+            deps.kotlin.stdlib,
+            deps.dagger.dagger
+    )
+    kapt(
+            deps.dagger.processor
     )
 }
