+++ features/notepad/build.gradle
@@ -43,6 +43,7 @@ android {
 
 dependencies {
     implementation project(":core")
+    implementation project(":base:notepad")
 
     implementation(
             deps.kotlin.stdlib,
