+++ features/notepad/src/main/kotlin/com/dvinc/notepad/domain/repository/note/NoteRepository.kt
@@ -7,7 +7,7 @@
 
 package com.dvinc.notepad.domain.repository.note
 
-import com.dvinc.notepad.domain.model.note.Note
+import com.dvinc.base.notepad.domain.model.Note
 import kotlinx.coroutines.flow.Flow
 
 interface NoteRepository {
