+++ features/notepad/src/main/kotlin/com/dvinc/notepad/domain/usecase/archive/ArchiveUseCase.kt
@@ -5,7 +5,7 @@
 
 package com.dvinc.notepad.domain.usecase.archive
 
-import com.dvinc.notepad.domain.model.note.Note
+import com.dvinc.base.notepad.domain.model.Note
 import com.dvinc.notepad.domain.repository.ArchiveRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
