+++ features/notepad/src/main/kotlin/com/dvinc/notepad/domain/usecase/notepad/NotepadUseCase.kt
@@ -7,7 +7,7 @@
 
 package com.dvinc.notepad.domain.usecase.notepad
 
-import com.dvinc.notepad.domain.model.note.Note
+import com.dvinc.base.notepad.domain.model.Note
 import com.dvinc.notepad.domain.repository.ArchiveRepository
 import com.dvinc.notepad.domain.repository.note.NoteRepository
 import kotlinx.coroutines.flow.Flow
