+++ features/notepad/src/main/kotlin/com/dvinc/notepad/presentation/adapter/notepad/NotepadSwipeCallback.kt
@@ -11,7 +11,7 @@ import android.graphics.drawable.ColorDrawable
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
 import com.dvinc.notepad.R
-import com.dvinc.notepad.presentation.model.NoteUi
+import com.dvinc.base.notepad.presentation.model.NoteUi
 
 class NotepadSwipeCallback(
     private val notesAdapter: NotepadAdapter,
