+++ features/notepad/src/main/kotlin/com/dvinc/notepad/presentation/ui/notepad/NotepadViewState.kt
@@ -6,7 +6,7 @@
 package com.dvinc.notepad.presentation.ui.notepad
 
 import com.dvinc.core.ui.ViewState
-import com.dvinc.notepad.presentation.model.NoteUi
+import com.dvinc.base.notepad.presentation.model.NoteUi
 
 data class NotepadViewState(
     val notes: List<NoteUi> = emptyList(),
