+++ features/notepad/src/test/kotlin/com/dvinc/notepad/data/repository/NoteDataRepositoryTest.kt
@@ -5,7 +5,7 @@ import com.dvinc.core.database.entity.note.NoteEntity
 import com.dvinc.notepad.BaseTest
 import com.dvinc.notepad.data.mapper.note.NoteDataMapper
 import com.dvinc.notepad.data.repository.note.NoteDataRepository
-import com.dvinc.notepad.domain.model.note.Note
+import com.dvinc.base.notepad.domain.model.Note
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
