+++ features/notepad/src/test/kotlin/com/dvinc/notepad/domain/usecase/NotepadUseCaseTest.kt
@@ -1,7 +1,6 @@
 package com.dvinc.notepad.domain.usecase
 
 import com.dvinc.notepad.BaseTest
-import com.dvinc.notepad.domain.model.note.Note
 import com.dvinc.notepad.domain.repository.ArchiveRepository
 import com.dvinc.notepad.domain.repository.note.NoteRepository
 import com.dvinc.notepad.domain.usecase.notepad.NotepadUseCase
